Rename functions.py to utils.py
