+++ CircleciScripts/run_integrationtest.py
@@ -1,4 +1,4 @@
-from functions import *
+from utils import *
 import sys
 import os
